requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Marshmallow==0.14.0
marshmallow-sqlalchemy==0.28.1
pytest==7.4.0
pytest-flask==1.2.0
Werkzeug<3.0

